requirements.txt
aiogram==2.11.2
aioschedule==0.5.2
beautifulsoup4==4.9.3
lxml==4.6.2
python-dotenv==0.15.0
requests==2.25.1
